+++ backend/internal/domain/repository/repository_mock/post_repository_mock.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -source=post_repository.go -destination=mock/post_repository_mock.go
+//	mockgen -source=post_repository.go -destination=repository_mock/post_repository_mock.go -package repository_mock
 //
 
-// Package mock_repository is a generated GoMock package.
-package mock_repository
+// Package repository_mock is a generated GoMock package.
+package repository_mock
 
 import (
 	context "context"
